\docType{package}
\name{rreporttools}
\alias{rreporttools}
\alias{rreporttools-package}
\title{A small set of tools for performing common reporting tasks}
\description{
rreporttools provides a small set of functions to make our
lives easier at the Division of Statistical Analysis and
Reporting. These functions extend several packages. We
extend data.table and plyr to include totals and subtotals.
We also extend xlsx to make formatting almost completely
automatic.
}

